Fix TwoFactorUsersAdminTest: load Patchwork before WP stubs
Brain\Monkey's Patchwork can only mock functions defined in files it
preprocessed via its stream wrapper. The bootstrap (tests/autoload.php)
defined WP stubs inline, after which Patchwork was loaded lazily inside
setUp() — too late to track those definitions.

Fix: load Patchwork explicitly at the top of the bootstrap so its stream
wrapper is active, then move all mockable WP stubs to tests/wp-stubs.php
(a separately-required file that Patchwork preprocesses). Also add
wp_cache_get/set/delete, get_site_url, and is_multisite stubs needed by
the template rendering path exercised by users_admin_section().
